refactor : scheduler 네이밍 변경
