+++ src/main/java/inu/codin/codin/domain/post/entity/PostEntity.java
@@ -4,8 +4,8 @@ import inu.codin.codin.common.dto.BaseTimeEntity;
 import inu.codin.codin.domain.post.dto.request.PostCreateRequestDTO;
 import inu.codin.codin.domain.post.exception.PostErrorCode;
 import inu.codin.codin.domain.post.exception.PostException;
-import inu.codin.codin.domain.post.schedular.exception.SchedulerErrorCode;
-import inu.codin.codin.domain.post.schedular.exception.SchedulerException;
+import inu.codin.codin.domain.post.scheduler.exception.SchedulerErrorCode;
+import inu.codin.codin.domain.post.scheduler.exception.SchedulerException;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.AccessLevel;
